+++ test/function/statistics/mean.test.js
@@ -75,6 +75,11 @@ describe('mean', function() {
     assert.throws(function() {mean([], 2, 3)});
   });
 
+  it('should throw an error if called a dimension out of range', function() {
+    assert.throws(function() {mean([1,2,3], -1)}, /IndexError: Index out of range \(-1 < 0\)/);
+    assert.throws(function() {mean([1,2,3], 1)}, /IndexError: Index out of range \(1 > 0\)/);
+  });
+
   it('should throw an error if called with an empty array', function() {
     assert.throws(function() {mean([])});
   });